stream colors [nfc]: Move StreamColorSwatch out to a new file in widgets/
Toward #393, "model: Track/memoize stream-color variants somewhere
other than /api/model".

But this only moves definitions; we'll reassign the
caching/computing logic in an upcoming commit.

Related: #393
