+++ test/api/model/model_checks.dart
@@ -1,5 +1,3 @@
-import 'dart:ui';
-
 import 'package:checks/checks.dart';
 import 'package:zulip/api/model/model.dart';
 
@@ -27,14 +25,6 @@ extension UserChecks on Subject<User> {
 
 extension ZulipStreamChecks on Subject<ZulipStream> {
   Subject<int?> get canRemoveSubscribersGroup => has((e) => e.canRemoveSubscribersGroup, 'canRemoveSubscribersGroup');
-}
-
-extension StreamColorSwatchChecks on Subject<StreamColorSwatch> {
-  Subject<Color> get base => has((s) => s.base, 'base');
-  Subject<Color> get unreadCountBadgeBackground => has((s) => s.unreadCountBadgeBackground, 'unreadCountBadgeBackground');
-  Subject<Color> get iconOnPlainBackground => has((s) => s.iconOnPlainBackground, 'iconOnPlainBackground');
-  Subject<Color> get iconOnBarBackground => has((s) => s.iconOnBarBackground, 'iconOnBarBackground');
-  Subject<Color> get barBackground => has((s) => s.barBackground, 'barBackground');
 }
 
 extension MessageChecks on Subject<Message> {
